renamed raven_embedded -> pyravendb_embedded
