+++ pyravendb_embedded/tools/helpers.py
